bridges: add a BROKEN Windows Arm64 bridge
This is a mixed port of the gcc3_linux_aarch64 bridge and the
refactored x86-64 exception handling. While the Aarch64 code
might work, the amd64 bytecode in the except.cxx definitely
won't.

This needs at least an update to the trampoline code, as
described in:
https://wiki.openoffice.org/wiki/Lazy_Hackers_Guide_To_Porting

It compiles, but will definitely just crash!

Change-Id: I9782a2e99c0231cdd1286af156ad312229eccf39